Installing the Post NPSP Configuration
The managed folder when installing dependency 9/9 was removed in pull request - https://github.com/SalesforceFoundation/Cumulus/pull/3900

It has been replaced with installing the post install configuration in the unpackaged folder. Line 75 includes a temporary workaround that gets around a NAMESPACE placeholder in the files. Opened up a pull request against the Cumulus repo so the workaround can be removed in the future - https://github.com/SalesforceFoundation/Cumulus/pull/4212